[platform] Move system EEPROM tests to separate file
